Implement reward shaping environment

+++ game/env_step_result.py
@@ -8,6 +8,6 @@ class EnvStepResult:
     """The result of a step in the game environment."""
 
     visible_board: list[list[int]]
-    reward: int
+    reward: float
     terminated: bool
     open_result: OpenResult
